+++ MASsenger.Application/Dtos/Read/BotReadDto.cs
@@ -2,7 +2,7 @@
 {
     public class BotReadDto
     {
-        public ulong Id { get; set; }
+        public Int32 Id { get; set; }
         public string Name { get; set; } = null!;
         public string? Username { get; set; }
         public string? Description { get; set; }
