+++ MASsenger.Application/Dtos/Read/MessageReadDto.cs
@@ -2,8 +2,8 @@
 {
     public class MessageReadDto
     {
-        public ulong SenderID { get; }
-        public ulong DestinationID { get; }
+        public Int32 SenderID { get; }
+        public Int32 DestinationID { get; }
         public string? Text { get; }
         public DateTime? SentTime { get; }
     }
