+++ MASsenger.Application/Dtos/Read/SystemMessageReadDto.cs
@@ -2,7 +2,7 @@
 {
     public class SystemMessageReadDto
     {
-        public ulong DestinationID { set; get; }
+        public Int32 DestinationID { set; get; }
         public string? Text { set; get; }
         public DateTime? SentTime { set; get; }
     }
