+++ MASsenger.Core/Entities/Message/BaseMessage.cs
@@ -1,10 +1,8 @@
-using MASsenger.Core.Enums;
-
 namespace MASsenger.Core.Entities
 {
     public class BaseMessage
     {
-        public UInt64 Id { get; set; } // id of the message inside of the group/chat/bot
+        public Int32 Id { get; set; } // id of the message inside of the chat/user
         public BaseChat Destination { get; set; } = null!;
         public DateTime SentTime { get; set; } = DateTime.Now;
